compar modal and styling
