+++ client/public/style.css
@@ -38,7 +38,7 @@ body {
   height: 400px;
   width: 904px;
   display: flex;
-  position: relative;
+
   align-items: center;
   justify-content: center;
   margin-left: 80px;
@@ -153,13 +153,18 @@ body {
 }
 
 .compare-modal-main {
-  position: absolute;
+  position: fixed;
   background: white;
+  border-radius: 4px;
+  border-color: red;
+  border-style: solid;
+  border-width: 2px;
   width: 80%;
   height: auto;
   top:50%;
   left: 50%;
   transform: translate(-50%, -50%);
+  overflow-y:scroll;
 
 }
 
@@ -168,6 +173,20 @@ body {
 }
 .compare-display-none {
   display: none;
+}
+
+table{
+margin:0 auto;
+}
+
+th{
+  font-size: 20px;
+  text-align: center;
+}
+
+td{
+  font-size: 15px;
+  text-align: center;
 }
