--- client/src/components/RelatedProducts/CompareTable.jsx
@@ -1,14 +1,0 @@
-import React from 'react';
-
-
-const CompareTable = (props) => {
- return (
-  <tbody>
-      <tr></tr>
-      <tr></tr>
-      <tr></tr>
-  </tbody>
- )
-}
-
-export default CompareTable
