+++ client/src/components/RelatedProducts/RelatedProducts.jsx
@@ -1,11 +1,15 @@
 import React, {useState, useEffect} from 'react';
 import RelatedProductsCards from './RelatedProductsCards.jsx';
 import YourOutfitCards from './YourOutfitCards.jsx'
+import CompareModal from './CompareModal.jsx'
 
 const RelatedProducts = (props) => {
   const [defaultPhotos, setDefaultPhotos] = useState({});
   const [photosLoaded, setPhotosLoaded] = useState(false)
   const [currentPhotoUrl, setCurrentPhotoUrl] = useState('');
+
+  const [showModal, setShowModal] = useState(false)
+  const [compareInfo, setCompareInfo] = useState([])
 
    useEffect(()=>{
     let photos={};
@@ -29,11 +33,23 @@ const RelatedProducts = (props) => {
      props.changeProducts(id)
    }
 
+   const starHandler = (current, related) => {
+
+    setCompareInfo([current, related])
+    setShowModal(true)
+
+ }
+
+ const closeModal = () => {
+   setShowModal(false)
+ }
+
   return(
   <div className="related-products-container">
 
     <div className="relatedproducts-title">Related Products</div>
     <RelatedProductsCards
+    modalhandler={starHandler}
     currentProduct={props.currentProduct}
     relatedProducts={props.relatedProducts}
     relatedProductStyles={props.relatedProductStyles}
@@ -42,6 +58,11 @@ const RelatedProducts = (props) => {
     relatedProductsPhotos={defaultPhotos}
     changeProducts={changeProducts}
     />
+
+  <CompareModal show={showModal} closemodal={closeModal}
+  compareinfo={compareInfo}
+  />
+
 
     <div className="youroutfit-title">Your Outfit</div>
     <YourOutfitCards
